refactor(core): check for updates once a day
