+++ packages/core/src/commands/update.ts
@@ -1,10 +1,11 @@
 import { list } from "../helpers/pm2";
+import { checkForUpdates } from "../helpers/update";
 import { BaseCommand } from "./command";
 
 export class UpdateCommand extends BaseCommand {
     public static description: string = "Update the core installation";
 
     public async run(): Promise<void> {
-        // const { flags } = this.parse(UpdateCommand);
+        await checkForUpdates(this.config);
     }
 }
